juntar y mover mesas.

+++ android/ValleTPV/ValleTPVLib/src/main/java/com/valleapp/valletpvlib/models/CuentaModel.kt
@@ -61,7 +61,7 @@ class CuentaModel(val mainModel: MainModel, private val camId: Long, private val
                 linea.camareroId = camId
                 linea.id = System.currentTimeMillis()
                 lineasDao.insert(linea)
-                mesasDao.abrirMesa(mesaId.toInt())
+                mesasDao.abrirMesa(mesaId)
             }
 
         }
